Updated for python3! For the first time in 4eva!
I opened this up expecting it to be an impossibly complicated project -
getting this to run - and instead discovered that it took about twenty
minutes of work, plus I pretty quickly organized things with a sub
library that I never got around to making before. NICE. Pretty dang
happy with that. Maybe more updates soon?